% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{ciIter}
\alias{ciIter}
\title{Function for determining confidence intervals}
\usage{
ciIter(gamma0, index, intercept, Y, X, U, epsilon, lambda, heteroskedastic,
  hc, kappa)
}
\arguments{
\item{gamma0}{Scalar, a conjectured value of the coefficient for
the covariate of interest, i.e. the covariate for which the
confidence interval is being constructed.}

\item{index}{Integer, indexes which covariate is of interest.}

\item{intercept}{Boolean, indicates whether an intercept term is
included in the regression.}

\item{Y}{Vector of dependent variable.}

\item{X}{Matrix of covariates.}

\item{U}{Vector of residuals from regressing \code{Y} on \code{X}
using l1svr.}

\item{epsilon}{Real scalar, bandwidth for SVR.}

\item{lambda}{Real scalar, parameter for SVR that scales the
penalty.}

\item{heteroskedastic}{Boolean, indicate whether data has
heteroskedastic errors. If set to \code{TRUE}, then density
estimation is performed when conducting inference.}

\item{hc}{Scalar determing the quantiles of a standard normal
distribution used to define the bandwidth for nonparametric
density estimation. See
\href{https://econpapers.repec.org/paper/attwimass/8818.htm}{Powell
(1991)} and
\href{https://yuehaob.github.io/assets/svml1.pdf}{Bai
et. al. (2021)} for details.}

\item{kappa}{Scalar that directly scales the bandwdith for
nonparametric density estimation. See
\href{https://econpapers.repec.org/paper/attwimass/8818.htm}{Powell
(1991)} and
\href{https://yuehaob.github.io/assets/svml1.pdf}{Bai
et. al. (2021)} for details.}

\item{target}{Target, equal to the size of the test.}
}
\description{
This function performs one iteration of estimating the confidence
interval. That is, it performs the l1-SVR regression rankscore test
under a single null hypothesis.
}
