% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{l1svr}
\alias{l1svr}
\title{Support vector regression and inference under l1 regularization}
\usage{
l1svr(formula, data, epsilon, lambda, h = NULL, kappa = NULL,
  solver = "gurobi", inference = TRUE, confidence.level = NULL,
  confidence.iter = 25, confidence.tol = 0.001,
  confidence.same = 1e-06)
}
\arguments{
\item{formula}{Regression formula, e.g., \code{y ~ x1 + x2}.}

\item{data}{Data frame.}

\item{epsilon}{Scalar defining soft-thresholding rule.}

\item{lambda}{Scalar parameter to scale l1 penalty.}

\item{h}{Scalar determining the quantiles of a standard normal
distribution used to define the bandwidth for nonparametric
density estimation. If no argument is provided, then errors are
assumed to be homoskedastic, and error density estimation is
not performed. See
\href{https://econpapers.repec.org/paper/attwimass/8818.htm}{Powell
(1991)} and
\href{https://yuehaob.github.io/assets/svml1.pdf}{Bai
et. al. (2021)} for details.}

\item{kappa}{Scalar that directly scales the bandwidth for
nonparametric density estimation. If no argument is provided,
then errors are assumed to be homoskedastic, and error density
estimation is not performed.  See
\href{https://econpapers.repec.org/paper/attwimass/8818.htm}{Powell
(1991)} and
\href{https://yuehaob.github.io/assets/svml1.pdf}{Bai
et. al. (2021)} for details.}

\item{solver}{Character, name of the linear programming package in
R used to obtain the bounds on the treatment effect. The
function supports \code{'gurobi'}, \code{'cplexapi'},
\code{'lpsolveapi'}. The name of the solver should be provided
with quotation marks.}

\item{inference}{Boolean, determines whether or not p-values and
confidence intervals will be estimated. Set to \code{TRUE} by
default.}

\item{confidence.level}{Scalar between 0 and 1. Determines the
confidence level for estimating confidence intervals.}

\item{confidence.iter}{Integer determining maximum number of
iterations performed when estimating the confidence interval.
Set to 25 by default.}

\item{confidence.tol}{Scalar between 0 and 1. Determines the
tolerance for the confidence level when estimating the
confidence interval.}

\item{confidence.same}{Scalar between 0 and 1. Sets the tolerance
for determining when the confidence interval estimate may no
longer be improved.}
}
\value{
A list containing the coefficient estimates from the
    regression, the p-values, and the confidence intervals.
}
\description{
This function performs the support vector regression under l1
regularization. A regression rankscore test is performed to
conduct inference and construct confidence intervals.
}
