% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{lpSetupSolver}
\alias{lpSetupSolver}
\title{Configure LP environment to be compatible with solvers}
\usage{
lpSetupSolver(model, solver)
}
\arguments{
\item{solver}{Character, the LP solver.}

\item{env}{List, the LP object.}
}
\value{
Nothing, as this modifies an environment variable to save
    memory.
}
\description{
This alters the LP object so the model will be compatible with
specific solvers.
}
