% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{print.l1svr}
\alias{print.l1svr}
\title{Print results}
\usage{
\method{print}{l1svr}(x, ...)
}
\arguments{
\item{x}{an object returned from '\code{l1svr}'.}

\item{...}{additional arguments.}
}
\value{
basic set of results.
}
\description{
This function uses the print method on the l1svr return list.
}
