% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Data}
\format{A data frame with 500 rows and 3 columns.
\describe{
  \item{y}{outcome variable}
  \item{x1}{covariate 1}
  \item{x2}{covariate 2}
}}
\source{
Simulated data.
}
\usage{
simdata
}
\description{
Simulated Data
}
\keyword{datasets}
