% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{summary.l1svr}
\alias{summary.l1svr}
\title{Summarize results}
\usage{
\method{summary}{l1svr}(object, ...)
}
\arguments{
\item{object}{an object returned from '\code{l1svr}'.}

\item{...}{additional arguments.}
}
\value{
summarized results.
}
\description{
This function uses the summary method on the l1svr return list.
}
