% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svr-functions.R
\name{svmRegress}
\alias{svmRegress}
\title{l1-SVR regression function}
\usage{
svmRegress(Y, X, epsilon, lambda, intercept = TRUE, solver)
}
\arguments{
\item{Y}{Vector, dependent variable.}

\item{X}{2d array, independent variables, excluding column of
constants for the intercept.}

\item{epsilon}{Parameter defining soft-thresholding rule.}

\item{lambda}{Parameter to scale l1 penalty.}

\item{intercept}{Boolean, set to \code{TRUE} if an intercept
should be included in the regression.}

\item{solver}{character, name of the linear programming package in
R used to obtain the bounds on the treatment effect. The
function supports \code{'gurobi'}, \code{'cplexapi'},
\code{'lpsolveapi'}. The name of the solver should be provided
with quotation marks.}
}
\value{
A list including the coefficient estimates, and
    solutions to the primal and dual problem that can be used to
    identify the support vectors and recover the coefficient
    estimates.
}
\description{
This function carries out an SVR regression under l1 penalty.
}
